pyproject.toml
[project]
name = "fastfinance"
version = "0.1.0"
description = "Personal net-worth manager to track finances, banks, subscriptions and investments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.22",
    "python-telegram-bot>=21.0",
    "paddleocr>=3.2.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
    "openpyxl>=3.1.5",
    "rich>=14.3.1",
    "sqlalchemy>=2.0.46",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
    "dotenv>=0.9.9",
    "pip>=26.0.1",
]

[project.scripts]
fastfinance = "patrimonio.cli:main"
fastfinance-web = "patrimonio.web.app:main"
fastfinance-bot = "patrimonio.telegram.bot:main"
fastfinance-mobile = "patrimonio.mobile.app:main"

[project.optional-dependencies]
mobile = [
    "kivy>=2.3.0",
    "kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip",
]

[tool.uv]
package = true
dev-dependencies = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
